requirements-base.txt
Django==1.5.4
Unipath==1.0
oauthlib==0.6.0
python-openid==2.2.5
python-social-auth==0.1.13
requests==2.0.0
requests-oauthlib==0.3.3
six==1.4.1
wsgiref==0.1.2
South==0.8.2
dj-static==0.0.5
django-autoslug==1.7.1
